+++ tests/data/cases/return_annotation_brackets_crash_line_length_1.py
@@ -1,0 +1,9 @@
+# flags: --preview --minimum-version=3.10 --line-length=1
+
+def foo() -> tuple[int, int,]:
+    ...
+# output
+def foo() -> tuple[
+    int,
+    int,
+]: ...